Unify interrupt and restart command handlers for IW and notebook
