+++ src/client/datascience/interactive-window/nativeInteractiveWindow.ts
@@ -3,7 +3,6 @@
 import type { nbformat } from '@jupyterlab/coreutils';
 import * as path from 'path';
 import {
-    ConfigurationTarget,
     Event,
     EventEmitter,
     NotebookCell,
@@ -34,7 +33,7 @@ import {
 import { JVSC_EXTENSION_ID, MARKDOWN_LANGUAGE, PYTHON_LANGUAGE } from '../../common/constants';
 import { ContextKey } from '../../common/contextKey';
 import '../../common/extensions';
-import { traceError, traceInfo } from '../../common/logger';
+import { traceInfo } from '../../common/logger';
 import { IFileSystem } from '../../common/platform/types';
 import * as uuid from 'uuid/v4';
 
@@ -46,7 +45,6 @@ import {
     Resource
 } from '../../common/types';
 import { createDeferred, Deferred } from '../../common/utils/async';
-import * as localize from '../../common/utils/localize';
 import { noop } from '../../common/utils/misc';
 import { generateCellsFromNotebookDocument } from '../cellFactory';
 import { CellMatcher } from '../cellMatcher';
@@ -57,7 +55,6 @@ import {
     InteractiveWindowMessages,
     ISubmitNewCell
 } from '../interactive-common/interactiveWindowTypes';
-import { JupyterKernelPromiseFailedError } from '../jupyter/kernels/jupyterKernelPromiseFailedError';
 import { IKernel, IKernelProvider, KernelConnectionMetadata } from '../jupyter/kernels/types';
 import { INotebookControllerManager } from '../notebook/types';
 import { VSCodeNotebookController } from '../notebook/vscodeNotebookController';
@@ -70,8 +67,6 @@ import {
     IInteractiveWindowLoadable,
     IJupyterDebugger,
     INotebookExporter,
-    InterruptResult,
-    IStatusProvider,
     WebViewViewChangeEventArgs
 } from '../types';
 import { createInteractiveIdentity, getInteractiveWindowTitle } from './identity';
@@ -115,7 +110,6 @@ export class NativeInteractiveWindow implements IInteractiveWindowLoadable {
     protected fileInKernel: string | undefined;
 
     private isDisposed = false;
-    private restartingKernel = false;
     private kernel: IKernel | undefined;
     private kernelLoadPromise: Promise<void> | undefined;
     private initialControllerSelected: Deferred<void>;
@@ -126,7 +120,6 @@ export class NativeInteractiveWindow implements IInteractiveWindowLoadable {
     constructor(
         private readonly applicationShell: IApplicationShell,
         private readonly documentManager: IDocumentManager,
-        private readonly statusProvider: IStatusProvider,
         private readonly fs: IFileSystem,
         private readonly configuration: IConfigurationService,
         private readonly commandManager: ICommandManager,
@@ -448,136 +441,6 @@ export class NativeInteractiveWindow implements IInteractiveWindowLoadable {
             }
         }
         return result;
-    }
-
-    // TODO Migrate all of this code into a common command handler
-    public async interruptKernel(): Promise<void> {
-        // trackKernelResourceInformation(this._notebook?.resource, { interruptKernel: true });
-        const notebookEditor = await this._editorReadyPromise;
-        if (this.kernel && !this.restartingKernel) {
-            const status = this.statusProvider.set(
-                localize.DataScience.interruptKernelStatus(),
-                true,
-                undefined,
-                undefined,
-                this
-            );
-
-            try {
-                const result = await this.kernel.interrupt(notebookEditor.document);
-                status.dispose();
-
-                // We timed out, ask the user if they want to restart instead.
-                if (result === InterruptResult.TimedOut && !this.restartingKernel) {
-                    const message = localize.DataScience.restartKernelAfterInterruptMessage();
-                    const yes = localize.DataScience.restartKernelMessageYes();
-                    const no = localize.DataScience.restartKernelMessageNo();
-                    const v = await this.applicationShell.showInformationMessage(message, { modal: true }, yes, no);
-                    if (v === yes) {
-                        await this.restartKernelInternal();
-                    }
-                } else if (result === InterruptResult.Restarted) {
-                    // Uh-oh, keyboard interrupt crashed the kernel.
-                    // this.addSysInfo(SysInfoReason.Interrupt).ignoreErrors(); // This should be handled in kernel.ts
-                }
-            } catch (err) {
-                status.dispose();
-                traceError(err);
-                this.applicationShell.showErrorMessage(err).then(noop, noop);
-            }
-        }
-    }
-
-    // TODO Migrate all of this code into a common command handler
-    public async restartKernel(): Promise<void> {
-        if (this.kernel && !this.restartingKernel) {
-            this.restartingKernel = true;
-            this.startProgress();
-
-            try {
-                if (await this.shouldAskForRestart()) {
-                    // Ask the user if they want us to restart or not.
-                    const message = localize.DataScience.restartKernelMessage();
-                    const yes = localize.DataScience.restartKernelMessageYes();
-                    const dontAskAgain = localize.DataScience.restartKernelMessageDontAskAgain();
-                    const no = localize.DataScience.restartKernelMessageNo();
-
-                    const v = await this.applicationShell.showInformationMessage(message, yes, dontAskAgain, no);
-                    if (v === dontAskAgain) {
-                        await this.disableAskForRestart();
-                        await this.restartKernelInternal();
-                    } else if (v === yes) {
-                        await this.restartKernelInternal();
-                    }
-                } else {
-                    await this.restartKernelInternal();
-                }
-            } finally {
-                this.restartingKernel = false;
-                this.stopProgress();
-            }
-        }
-    }
-
-    private async shouldAskForRestart(): Promise<boolean> {
-        const settings = this.configuration.getSettings(this.owningResource);
-        return settings && settings.askForKernelRestart === true;
-    }
-
-    private async disableAskForRestart(): Promise<void> {
-        const settings = this.configuration.getSettings(this.owningResource);
-        if (settings) {
-            this.configuration
-                .updateSetting('askForKernelRestart', false, undefined, ConfigurationTarget.Global)
-                .ignoreErrors();
-        }
-    }
-
-    private async restartKernelInternal(): Promise<void> {
-        this.restartingKernel = true;
-        const notebookEditor = await this._editorReadyPromise;
-
-        // Set our status
-        const status = this.statusProvider.set(
-            localize.DataScience.restartingKernelStatus(),
-            true,
-            undefined,
-            undefined,
-            this
-        );
-
-        try {
-            if (this.kernel && notebookEditor) {
-                await this.kernel.restart(notebookEditor.document);
-
-                // Reset our file in the kernel.
-                const fileInKernel = this.fileInKernel;
-                this.fileInKernel = undefined;
-                if (fileInKernel) {
-                    // TODO this should really be done in the IKernel itself
-                    await this.setFileInKernel(fileInKernel, notebookEditor.document);
-                }
-
-                // // Compute if dark or not.
-                // const knownDark = await this.isDark();
-
-                // // Before we run any cells, update the dark setting
-                // await this.kernel?.notebook?.setMatplotLibStyle(knownDark);
-            }
-        } catch (exc) {
-            // If we get a kernel promise failure, then restarting timed out. Just shutdown and restart the entire server
-            if (exc instanceof JupyterKernelPromiseFailedError && this.kernel) {
-                await this.kernel.dispose();
-                await this.kernel.restart(notebookEditor.document);
-            } else {
-                // Show the error message
-                this.applicationShell.showErrorMessage(exc).then(noop, noop);
-                traceError(exc);
-            }
-        } finally {
-            status.dispose();
-            this.restartingKernel = false;
-        }
     }
 
     public undoCells() {
