+++ src/client/datascience/notebook/notebookEditor.ts
@@ -4,37 +4,24 @@
 'use strict';
 
 import {
-    ConfigurationTarget,
     Event,
     EventEmitter,
     NotebookCellKind,
     NotebookRange,
     NotebookDocument,
-    ProgressLocation,
     Uri,
     NotebookCellData,
     NotebookCell,
     NotebookData
 } from 'vscode';
-import { IApplicationShell, ICommandManager, IVSCodeNotebook } from '../../common/application/types';
-import { traceError, traceInfo } from '../../common/logger';
-import { IConfigurationService, IDisposable, IDisposableRegistry, IExtensions } from '../../common/types';
-import { DataScience } from '../../common/utils/localize';
+import { ICommandManager, IVSCodeNotebook } from '../../common/application/types';
+import { IDisposable, IDisposableRegistry, IExtensions } from '../../common/types';
 import { isUntitledFile, noop } from '../../common/utils/misc';
-import { StopWatch } from '../../common/utils/stopWatch';
-import { captureTelemetry, sendTelemetryEvent } from '../../telemetry';
+import { captureTelemetry } from '../../telemetry';
 import { Telemetry } from '../constants';
-import { sendKernelTelemetryEvent, trackKernelResourceInformation } from '../telemetry/telemetry';
-import { JupyterKernelPromiseFailedError } from '../jupyter/kernels/jupyterKernelPromiseFailedError';
-import { IKernel, IKernelProvider } from '../jupyter/kernels/types';
 import {
     INotebook,
-    INotebookEditor,
-    INotebookExecutionLogger,
-    INotebookProvider,
-    InterruptResult,
-    IStatusProvider
-} from '../types';
+    INotebookEditor} from '../types';
 import { NotebookCellLanguageService } from './cellLanguageService';
 import { chainWithPendingUpdates } from './helpers/notebookUpdater';
 import { getNotebookMetadata } from './helpers/helpers';
@@ -64,20 +51,12 @@ export class NotebookEditor implements INotebookEditor {
     private _closed = new EventEmitter<INotebookEditor>();
     private _saved = new EventEmitter<INotebookEditor>();
     private _modified = new EventEmitter<INotebookEditor>();
-    private restartingKernel?: boolean;
-    private kernelInterruptedDontAskToRestart: boolean = false;
     constructor(
         public readonly document: NotebookDocument,
         private readonly vscodeNotebook: IVSCodeNotebook,
         private readonly commandManager: ICommandManager,
-        private readonly notebookProvider: INotebookProvider,
-        private readonly kernelProvider: IKernelProvider,
-        private readonly statusProvider: IStatusProvider,
-        private readonly applicationShell: IApplicationShell,
-        private readonly configurationService: IConfigurationService,
         disposables: IDisposableRegistry,
         private readonly cellLanguageService: NotebookCellLanguageService,
-        private loggers: INotebookExecutionLogger[],
         private extensions: IExtensions
     ) {
         vscodeNotebook.onDidCloseNotebookDocument(this.onClosedDocument, this, disposables);
@@ -207,83 +186,7 @@ export class NotebookEditor implements INotebookEditor {
             }).then(noop, noop);
         }
     }
-    public async interruptKernel(): Promise<void> {
-        if (this.restartingKernel) {
-            traceInfo(`Interrupt requested & currently restarting ${this.document.uri} in notebookEditor.`);
-            trackKernelResourceInformation(this.document.uri, { interruptKernel: true });
-            return;
-        }
-        const kernel = this.kernelProvider.get(this.document);
-        if (!kernel || this.restartingKernel) {
-            traceInfo(
-                `Interrupt requested & no kernel or currently restarting ${this.document.uri} in notebookEditor.`
-            );
-            trackKernelResourceInformation(this.document.uri, { interruptKernel: true });
-            return;
-        }
-        const status = this.statusProvider.set(DataScience.interruptKernelStatus(), true, undefined, undefined);
 
-        try {
-            traceInfo(`Interrupt requested & sent for ${this.document.uri} in notebookEditor.`);
-            const result = await kernel.interrupt(this.document);
-            if (result === InterruptResult.TimedOut) {
-                const message = DataScience.restartKernelAfterInterruptMessage();
-                const yes = DataScience.restartKernelMessageYes();
-                const no = DataScience.restartKernelMessageNo();
-                const v = await this.applicationShell.showInformationMessage(message, { modal: true }, yes, no);
-                if (v === yes) {
-                    this.restartingKernel = false;
-                    this.kernelInterruptedDontAskToRestart = true;
-                    await this.restartKernel();
-                }
-            }
-        } catch (err) {
-            traceError('Failed to interrupt kernel', err);
-            void this.applicationShell.showErrorMessage(err);
-        } finally {
-            this.kernelInterruptedDontAskToRestart = false;
-            status.dispose();
-        }
-    }
-
-    public async restartKernel(): Promise<void> {
-        trackKernelResourceInformation(this.document.uri, { restartKernel: true });
-        sendTelemetryEvent(Telemetry.RestartKernelCommand);
-        if (this.restartingKernel) {
-            trackKernelResourceInformation(this.document.uri, { restartKernel: true });
-            return;
-        }
-        const kernel = this.kernelProvider.get(this.document);
-
-        if (kernel && !this.restartingKernel) {
-            if (await this.shouldAskForRestart()) {
-                // Ask the user if they want us to restart or not.
-                const message = DataScience.restartKernelMessage();
-                const yes = DataScience.restartKernelMessageYes();
-                const dontAskAgain = DataScience.restartKernelMessageDontAskAgain();
-                const no = DataScience.restartKernelMessageNo();
-
-                const response = await this.applicationShell.showInformationMessage(message, yes, dontAskAgain, no);
-                if (response === dontAskAgain) {
-                    await this.disableAskForRestart();
-                    void this.applicationShell.withProgress(
-                        { location: ProgressLocation.Notification, title: DataScience.restartingKernelStatus() },
-                        () => this.restartKernelInternal(kernel)
-                    );
-                } else if (response === yes) {
-                    void this.applicationShell.withProgress(
-                        { location: ProgressLocation.Notification, title: DataScience.restartingKernelStatus() },
-                        () => this.restartKernelInternal(kernel)
-                    );
-                }
-            } else {
-                void this.applicationShell.withProgress(
-                    { location: ProgressLocation.Notification, title: DataScience.restartingKernelStatus() },
-                    () => this.restartKernelInternal(kernel)
-                );
-            }
-        }
-    }
     public dispose() {
         this._closed.fire(this);
     }
@@ -291,70 +194,6 @@ export class NotebookEditor implements INotebookEditor {
     private onClosedDocument(e?: NotebookDocument) {
         if (this.document === e) {
             this._closed.fire(this);
-        }
-    }
-
-    private async restartKernelInternal(kernel: IKernel): Promise<void> {
-        this.restartingKernel = true;
-
-        // Set our status
-        const status = this.statusProvider.set(DataScience.restartingKernelStatus(), true, undefined, undefined);
-
-        const stopWatch = new StopWatch();
-        try {
-            await kernel.restart(this.document);
-            sendKernelTelemetryEvent(this.document.uri, Telemetry.NotebookRestart, stopWatch.elapsedTime);
-        } catch (exc) {
-            // If we get a kernel promise failure, then restarting timed out. Just shutdown and restart the entire server.
-            // Note, this code might not be necessary, as such an error is thrown only when interrupting a kernel times out.
-            sendKernelTelemetryEvent(
-                this.document.uri,
-                Telemetry.NotebookRestart,
-                stopWatch.elapsedTime,
-                undefined,
-                exc
-            );
-            if (exc instanceof JupyterKernelPromiseFailedError && kernel) {
-                // Old approach (INotebook is not exposed in IKernel, and INotebook will eventually go away).
-                const notebook = await this.notebookProvider.getOrCreateNotebook({
-                    resource: this.file,
-                    identity: this.file,
-                    getOnly: true
-                });
-                if (notebook) {
-                    await notebook.dispose();
-                }
-                await this.notebookProvider.connect({
-                    getOnly: false,
-                    disableUI: false,
-                    resource: this.file,
-                    metadata: getNotebookMetadata(this.document)
-                });
-            } else {
-                // Show the error message
-                void this.applicationShell.showErrorMessage(exc);
-                traceError(exc);
-            }
-        } finally {
-            status.dispose();
-            this.restartingKernel = false;
-            this.loggers.forEach((l) => l.onKernelRestarted(this.file));
-        }
-    }
-    private async shouldAskForRestart(): Promise<boolean> {
-        if (this.kernelInterruptedDontAskToRestart) {
-            return false;
-        }
-        const settings = this.configurationService.getSettings(this.file);
-        return settings && settings.askForKernelRestart === true;
-    }
-
-    private async disableAskForRestart(): Promise<void> {
-        const settings = this.configurationService.getSettings(this.file);
-        if (settings) {
-            this.configurationService
-                .updateSetting('askForKernelRestart', false, undefined, ConfigurationTarget.Global)
-                .ignoreErrors();
         }
     }
 }
